change int32 to uint64

+++ sort.go
@@ -5,9 +5,9 @@ import (
 )
 
 func Sort(s Slice) {
-	sortutil.Int32Slice(s).Sort()
+	sortutil.Uint64Slice(s).Sort()
 }
 
 func (s Slice) Sort() {
-	sortutil.Int32Slice(s).Sort()
+	sortutil.Uint64Slice(s).Sort()
 }
